+++ tests/test_main.py
@@ -1,8 +1,7 @@
-from currenapp.main import app, is_validargs
+from currenapp.main import app
 from typer.testing import CliRunner
+from httmock import HTTMock, all_requests
 import pytest
-# from httmock import urlmatch, HTTMock
-import json
 
 
 runner = CliRunner()
@@ -11,17 +10,6 @@ runner = CliRunner()
 @pytest.fixture
 def cmd_args():
     return ["--base", "USD", "--target", "BRL"]
-
-
-def test_args_validation_wth_wrong_inputs():
-    mock_args = (("1b", "_BR"), ("abcd", "-az"), ("142", "!aW"))
-    for margs in mock_args:
-        assert not is_validargs(*margs)
-
-
-def test_args_validation_wth_right_inputs():
-    mock_args = ("ABC", "DEF")
-    assert is_validargs(*mock_args)
 
 
 def test_cli_error_wrong_currency_values(cmd_args):
@@ -41,7 +29,32 @@ def test_input_not_a_number(cmd_args):
     assert result.exit_code == 1
     assert "Expected a number" in result.stdout
 
-# Maybe useful
-# @urlmatch(netloc=r"(.*\.)?exchangerate-api")
-# def api_mock(url, request):
-    # return json.dumps({"result": "sucess"})
+
+@all_requests
+def api_mock(url, request):
+    return {
+        "status_code": 200,
+        "content": {
+            "result": "success",
+            "documentation": "https://www.exchangerate-api.com/docs",
+            "terms_of_use": "https://www.exchangerate-api.com/terms",
+            "time_last_update_unix": 1585267200,
+            "time_last_update_utc": "Fri, 27 Mar 2020 00:00:00 +0000",
+            "time_next_update_unix": 1585270800,
+            "time_next_update_utc": "Sat, 28 Mar 2020 01:00:00 +0000",
+            "base_code": "EUR",
+            "target_code": "GBP",
+            "conversion_rate": 0.8,
+            "conversion_result": 5.80,
+        },
+    }
+
+
+# FULL test of the application
+def main():
+    with HTTMock(api_mock):
+        app()
+
+
+if __name__ == "__main__":
+    main()